+++ Assets/Scripts/SideBorderCollision.cs
@@ -5,6 +5,8 @@ using UnityEngine;
 public class SideBorderCollision : MonoBehaviour
 {
     private Game _game;
+
+    private Forcefield _forcefield;
     
     // Which side this border is on
     [SerializeField] private int side;
@@ -13,12 +15,17 @@ public class SideBorderCollision : MonoBehaviour
     void Awake()
     {
         _game = GameObject.FindGameObjectWithTag("GameController").GetComponent<Game>();
+
+        _forcefield = GetComponent<Forcefield>();
     }
     
     void OnTriggerEnter2D(Collider2D collider)
     {
         Debug.Log("Side border " + side + " hit by ball");
-        
-        _game.OnBorderHit(side);
+
+        if (!_forcefield.IsActive)
+        {
+            _game.OnBorderHit(side);
+        }
     }
 }