+++ Assets/Scripts/powerups/PaddleShrinkPowerup.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 public class PaddleShrinkPowerup : Powerup
 {
     
-    protected override void OnHit(Transform ball)
+    protected override void OnHit(Transform ball, Transform paddle)
     {
-        base.OnHit(ball);
+        base.OnHit(ball, paddle);
 
         BallPaddleShrink ballPaddleShrink = ball.GetComponent<BallPaddleShrink>();